### 2025-03-10 #7738 @morrowcj
> [@morrowcj](https://github.com/morrowcj) Do you have tests?

Yes, tests can be found in at [tests/testthat/](https://github.com/morrowcj/remotePARTS/tree/joss_submission/tests/testthat).
