### 2025-03-19 #7738 @iosefa
> 👋 [@iosefa](https://github.com/iosefa) - would you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html

@jGaboardi Yes, I would be happy to review it.
