### 2025-03-20 #7738 @jGaboardi
> > 👋 [@iosefa](https://github.com/iosefa) - would you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
>
> [@jGaboardi](https://github.com/jGaboardi) Yes, I would be happy to review it.

Great! Thanks!
