### 2025-03-20 #7738 @morrowcj
Odd. Here is [Chamini's JOSS Reviewer's page](https://reviewers.joss.theoj.org/reviewers/3745). The GitHub handle listed is `katekanecm`, but maybe it is no longer active.
