### 2025-03-21 #7738 @karlosmantilla
> 👋 [@karlosmantilla](https://github.com/karlosmantilla) - would you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html

👋 Hi @jGaboardi, thank you for the invitation — I'm available and willing to review the submission for JOSS. I’ll make sure to follow the checklist and the review guidelines provided, and I’ll be attentive to the next communication to begin the review. Greetings!
