### 2025-03-21 #7738 @jGaboardi
> > 👋 [@karlosmantilla](https://github.com/karlosmantilla) - would you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
>
> 👋 Hi [@jGaboardi](https://github.com/jGaboardi), thank you for the invitation — I'm available and willing to review the submission for JOSS. I’ll make sure to follow the checklist and the review guidelines provided. Greetings!

Great! Thanks!
