### 2025-03-21 #7738 @highamm
> 👋 [@highamm](https://github.com/highamm) - would you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html

Hi @jGaboardi , Apologies for the late reply (it's our spring break this week!), but I am booked until the end of our semester in mid-May (Though if I followed the thread, you might have found your two reviewers for this anyway). Thanks for the invite, and I look forward to reviewing for JOSS again in the future!
