### 2025-03-21 #7937 @editorialbot
[REVIEW]: remotePARTS: Spatiotemporal autoregression analyses for large data sets
**Submitting author:** @morrowcj (<a href="http://orcid.org/0000-0002-3069-3296">Clay J. Morrow</a>)
**Repository:** https://github.com/morrowcj/remotePARTS
**Branch with paper.md** (empty if default branch): joss_submission
**Version:** v1.0.4.2
