### 2025-04-11 #7937 @karlosmantilla
Hi @morrowcj,

Thanks for your submission. I've completed my initial review of the `remotePARTS` package and the associated JOSS paper. The software installs and runs as expected, and it clearly addresses a need for partitioned spatio-temporal modeling using remote sensing data. The documentation is solid, and the paper is clearly written and well-structured.

Before recommending acceptance, I’d suggest addressing a couple of points that could improve the overall robustness and clarity of the project:

- Add basic automated tests to support reproducibility and increase reliability.
- Expand a bit on the internal function structure or general software architecture - this would help future users and developers.
- Two contributors are listed in the repository, but authorship details are not fully clear in the paper. If multiple authors are involved, clarifying individual contributions using CRediT roles would be helpful.

Let me know if you have any questions - happy to follow up as needed. Great work so far!
