### 2025-04-11 #7937 @morrowcj
Hello, @karlosmantilla,

Thanks for looking over the code. I've responded to a few of your comments directly below.

> Before recommending acceptance, I’d suggest addressing a couple of points that could improve the overall robustness and clarity of the project:
>
>     * Add basic automated tests to support reproducibility and increase reliability.

Could you provide more specific feedback regarding the tests? Did you not find them? If not, they are in the `tests/testthat/` directory. You may have been looking in the wrong branch. I had to add them for this submission, so they are only present in the `joss_submission` branch.

If it is that the tests that are present are insufficient, I welcome feedback about what to change or add!

>     * Expand a bit on the internal function structure or general software architecture - this would help future users and developers.

I'm not quite sure what you mean by this. Could you elaborate? Do you mean that I need to better describe how the code works? If so, what additional documentation do you have in mind?

>     * Two contributors are listed in the repository, but authorship details are not fully clear in the paper. If multiple authors are involved, clarifying individual contributions using CRediT roles would be helpful.

Ah, yes. My coauthor was responsible for the theoretical method development (Conceptualization, Funding Acquisition, Methodology, Writing - Review and Editing) and my contribution has largely been implementation, testing, etc. (Project Administration, Software, Validation, Visualization, Writing - Original Draft). Where is the best location to document this? Thanks for directing me to the CRediT framework - I'd not heard of it before.
