### 2025-04-11 #7937 @morrowcj
I should also add that each of the exported functions have extensive `@examples` that are executed each time a push or pull request is initiated (via `R CMD check` GitHub actions). This has been one way that I ensure the code works as expected, in addition to the tests. I think that `R CMD check` also executes the automated tests in its pipeline.
