### 2025-04-11 300b25f
Update check-standard.yaml
add the `joss_submission` branch to the github actions
