+++ .github/workflows/check-standard.yaml
@@ -2,9 +2,9 @@
 ## https://github.com/r-lib/actions/blob/v2/examples/check-standard.yaml
 on:
   push:
-    branches: [develop, master]
+    branches: [develop, master, joss_submission]
   pull_request:
-    branches: [develop, master]
+    branches: [develop, master, joss_submission]
 
 name: R-CMD-check
 
