### 2025-04-11 #7937 @karlosmantilla
Hi @morrowcj,

Thanks again for your detailed response, much appreciated! Here's my follow-up based on your clarifications, the contents of the `joss_submission` branch, and the recent updates to the repository:

### Automated tests & examples

You're absolutely right, I initially reviewed the `main` branch and missed the tests in `joss_submission`. I've now checked the `tests/testthat/` directory and confirmed that the key functions are covered.
Also, thanks for pointing out that the exported functions include runnable `@examples`, and that these are executed via `R CMD check` in GitHub Actions. I’ve seen your recent update in [PR #31](https://github.com/morrowcj/remotePARTS/pull/31), which adds the `joss_submission` branch to the CI workflow, this effectively resolves the issue of automated checks not running in the review branch. Great work on that!
No further action needed here, the setup now supports reproducibility and reliability.

### Software architecture / internal flow

To clarify: I was referring to a brief explanation of the *typical workflow* when using the package, for example:

- What are the key functions a user should start with?
- In what order are they generally used?
- How does a user go from loading data to running models and interpreting the results?

A short overview like “load data → run partitioned regressions → combine results”, placed in the vignette, README, or documentation site, would go a long way in helping new users understand how to structure their use of the package. Optionally, a small diagram could reinforce this visually.

The goal is to show how the different components of the package work together, especially for users unfamiliar with the underlying PARTS methodology.

### Author Contributions (CRediT roles)

Both authors are correctly listed in the `paper.md` with full affiliations, that's great. To align fully with JOSS guidelines, I suggest adding a short section titled **Author Contributions** using [CRediT roles](https://credit.niso.org/), describing the contributions of each author.

You can place this section right before the **Acknowledgments**, like this:

```markdown
# Author Contributions

Clay J. Morrow: Software, Validation, Visualization, Writing – Original Draft, Project Administration.
Anthony R. Ives: Conceptualization, Methodology, Funding Acquisition, Writing – Review & Editing.
```
