### 2025-04-11 #7937 @morrowcj
@karlosmantilla,

I appreciate the quick and thorough feedback.

> ### Software architecture / internal flow
>
> To clarify: I was referring to a brief explanation of the _typical workflow_ when using the package, for example:

Ah! I understand now. Have you looked at [Alaska vignette](https://morrowcj.github.io/remotePARTS/Alaska.html) referenced in the README file? This document goes through a typical workflow of how to use the package. The version published on my GitHub Pages (linked above) is tied to the `main` branch, but you should be able to access the newest version with

```r
remotes::install_github("morrowcj/remotePARTS", ref = "devel", build_vignettes = TRUE)
library(remotePARTS)
vignette("Alaska")
```

Should I provide a (much) shorter explanation within the README directly?

> ### Author Contributions (CRediT roles)

I will make a PR to add the author contributions. Thanks!
