### 2025-04-11 morrowcj/remotePARTS#33 @morrowcj
README needs updating with better instructions.
As brought to my attention via JOSS review (openjournals/joss-reviews#7937), the package does not have "clear guidelines for third parties wishing to 1. Contribute to the software 2. Report issues or problems with the software 3. Seek support".

* [ ] These should be added to the README file.
* [ ] Additionally, simple instructions for a typical use case workflow should be added to the README. Pointing to the vignette is perhaps not sufficient.
* [ ] I'd also like to add the "in review" badge to the README:
