### 2025-04-11 #7937 @karlosmantilla
Hi @morrowcj,

Thanks for your clarifications. I’ve reviewed the current state of the repository and the documentation, and I just have one minor suggestion for possible improvement.

### Software architecture / internal flow

The *Alaska* vignette provides a complete and clear demonstration of the typical workflow when using the package. That said, you might consider adding a brief summary of that workflow directly in the `README`.

There are two places where this could be helpful:
1. **After the `## Description` section** (as a short "Typical workflow" section outlining how the main components fit together).
2. **Within the `## Installation` section** (adding instructions on how to install the development version with `build_vignettes = TRUE` so users can access the vignette right away).

While not required, this addition could help reduce the learning curve for new users, especially those unfamiliar with the PARTS methodology or with how R handles vignettes.

Here is a possible example for the `README`:

```markdown
## Quick Overview

`remotePARTS` is designed for partitioned spatiotemporal modeling of large remote-sensing datasets. A typical workflow involves:

1. Preparing your dataset in wide format (e.g., one row per pixel).
2. Running partitioned regressions using built-in modeling functions.
3. Aggregating model outputs across partitions to perform inference.

For a complete example, see the [Alaska vignette](https://morrowcj.github.io/remotePARTS/Alaska.html).

To install the latest version with vignettes:
> ```r
> remotes::install_github("morrowcj/remotePARTS", ref = "devel", build_vignettes = TRUE)
> ```
```
