### 2025-04-11 71a2a23
Initial refresh on README
updated the badges and some language of the readme. This was done as part of the JOSS review process. Additional work on this file is still needed.
