+++ README.Rmd
@@ -13,56 +13,64 @@ knitr::opts_chunk$set(echo = TRUE, collapse = TRUE, comment = "#>",
 # remotePARTS
 
 <!-- badges: start -->
-[![License: GPL v3](https://img.shields.io/badge/License-GPLv3-blue.svg)](https://www.gnu.org/licenses/gpl-3.0)
-
-<!-- [![Lifecycle: maturing](https://img.shields.io/badge/lifecycle-maturing-blue.svg)](https://www.tidyverse.org/lifecycle/#maturing) -->
-
 [![Lifecycle: stable](https://img.shields.io/badge/lifecycle-stable-green.svg)](https://lifecycle.r-lib.org/articles/stages.html#stable)
-
-<!-- [![Travis build status](https://travis-ci.com/morrowcj/remotePARTS.svg?branch=master)](https://travis-ci.com/morrowcj/remotePARTS) -->
-
-<!-- [![Travis build status](https://travis-ci.com/morrowcj/remotePARTS.svg?branch=master)](https://travis-ci.org/github/morrowcj/remotePARTS) -->
-
-
+[![License: GPL v3](https://img.shields.io/badge/License-GPLv3-blue.svg)](https://www.gnu.org/licenses/gpl-3.0)
 [![R-CMD-check](https://github.com/morrowcj/remotePARTS/workflows/R-CMD-check/badge.svg)](https://github.com/morrowcj/remotePARTS/actions)
+[![status](https://joss.theoj.org/papers/c6a3da6a56aa0fb0e1f8a4f36cab12c2/status.svg)](https://joss.theoj.org/papers/c6a3da6a56aa0fb0e1f8a4f36cab12c2)
+
 <!-- badges: end -->
 
-`remotePARTS` is an `R` package that contains tools for analyzing spatiotemporal
-data, typically obtained via remote sensing. 
+`remotePARTS` is an software package for the `R` statistical programming language. 
+The package contains tools for analyzing spatiotemporal data, typically obtained 
+via remote sensing.
 
 ## Description
 
 These tools were created to test map-scale hypotheses about trends in large 
-remotely sensed data sets but any data with spatial and temporal variation
-can be analyzed. Tests are conducted using the PARTS method for analyzing spatially
-autocorrelated time series (Ives et al., 2021). The method's unique approach can
-handle extremely large data sets that other spatiotemporal models cannot, while
-still appropriately accounting for spatial and temporal autocorrelation. This 
-is done by partitioning the data into smaller chunks, analyzing chunks separately
-and then combining the separate analyses into a single, correlated test of the 
-map-scale hypotheses.
+remotely sensed data sets, but they are useful for analyzing trends in any 
+spatial data, with or without a temporal compoenent. Statistical tests are 
+conducted with the PARTS method for analyzing spatially autocorrelated time 
+series (Ives et al., 2021). The method's unique approach can handle extremely
+large data sets that other spatiotemporal models cannot, while still 
+appropriately accounting for spatial and temporal autocorrelation. This is 
+done by partitioning the data into smaller chunks, analyzing chunks separately
+and then combining the separate analyses into a single test, that acconts for
+correlations among chunks, of the map-scale hypotheses.
 
 ## Instalation
 
-To install the package and it's dependencies, use the following R code:
+To install the package and it's dependencies from CRAN, use the following R code:
 
 ```{r, eval = FALSE}
 install.packages("remotePARTS")
 ```
 
-To install the latest development version of this package from github, use
+To install the latest stable version of this package from github, use
 
 ```{r, eval = FALSE}
-install.packages("devtools") # ensure you have the latest devtools
-devtools::install_github("morrowcj/remotePARTS")
+remotes::install_github("morrowcj/remotePARTS")
 ```
 
[…]
