+++ README.Rmd
@@ -6,11 +6,11 @@ output: github_document
 
 ```{r, include = FALSE}
 knitr::opts_chunk$set(echo = TRUE, collapse = TRUE, comment = "#>",
-  fig.path = "man/figures/README-", out.width = "100%"
+                      fig.path = "man/figures/README-", out.width = "100%"
 )
 ```
 
-# remotePARTS
+## remotePARTS
 
 <!-- badges: start -->
 [![Lifecycle: stable](https://img.shields.io/badge/lifecycle-stable-green.svg)](https://lifecycle.r-lib.org/articles/stages.html#stable)
@@ -20,7 +20,7 @@ knitr::opts_chunk$set(echo = TRUE, collapse = TRUE, comment = "#>",
 
 <!-- badges: end -->
 
-`remotePARTS` is an software package for the `R` statistical programming language. 
+*remotePARTS* is an software package for the *R* statistical programming language. 
 The package contains tools for analyzing spatiotemporal data, typically obtained 
 via remote sensing.
 
@@ -28,13 +28,13 @@ via remote sensing.
 
 These tools were created to test map-scale hypotheses about trends in large 
 remotely sensed data sets, but they are useful for analyzing trends in any 
-spatial data, with or without a temporal compoenent. Statistical tests are 
+spatial data, with or without a temporal component. Statistical tests are 
 conducted with the PARTS method for analyzing spatially autocorrelated time 
 series (Ives et al., 2021). The method's unique approach can handle extremely
 large data sets that other spatiotemporal models cannot, while still 
-appropriately accounting for spatial and temporal autocorrelation. This is 
+appropriately accounting for autocorrelation structure. This is 
 done by partitioning the data into smaller chunks, analyzing chunks separately
-and then combining the separate analyses into a single test, that acconts for
+and then combining the separate analyses into a single test, that accounts for
 correlations among chunks, of the map-scale hypotheses.
 
 ## Instalation
@@ -56,41 +56,25 @@ and to test out the newest features and functionality, use
 ```{r, eval = FALSE}
 remotes::install_github("morrowcj/remotePARTS", ref = "develop")
 ```
-Then, upon successful installation, load the package with `library(remotePARTS)`. 
+
+To ensure the vignette is built when installing from GitHub, use 
+
+```{r, eval = FALSE}
+remotes::install_github("morrowcj/remotePARTS", build_vignettes = TRUE)
+```
+
+Then, upon successful installation, load the package with 
+`library(remotePARTS)`. 
 
 ### Dependencies
 
-Since the matrix operations in this package rely on `C++` code, as implemented 
+Since the matrix operations in this package rely on C++ code, as implemented 
 via the [RcppEigen package](https://github.com/RcppCore/RcppEigen),
 the latest version of [Rtools](https://cran.r-project.org/bin/windows/Rtools/)
 is required for Windows and C++11 is required for other systems. 
 
 <!--- Now that the package is on CRAN, I am not 100% certain that the above statement
 is strictly true anymore, but I'm going to leave it for now until I learn more. --->
-
-## Example usage
-
-<!-- TODO: This whole section needs to be reworked --->
-
-For examples on how to use `remotePARTS`, see the `Alaska` vignette:
-
-```{r, eval = FALSE}
-vignette("Alaska")
-```
-
-Note that the vignette needs to be built when installing with and may require 
[…]
