### 2025-04-12 #7937 @morrowcj
@karlosmantilla,

I have updated the README file (in the [morrowcj-patch-1 branch](https://github.com/morrowcj/remotePARTS/tree/morrowcj-patch-1)) for the morrowcj/remotePARTS#32 pull request.

Would you please look it over and let me know if it is 1) brief enough, 2) provides a sufficient summary of the typical use case and, 3) makes sense?

Thanks!
