### 2025-04-13 #7937 @karlosmantilla
Hi @morrowcj, @jGaboardi,

I’ve reviewed the latest updates, including the recent changes to the `README` and the finalized paper. Everything appears complete and consistent with the JOSS review criteria.

At this point, I have no further comments and consider the submission ready for acceptance.
Thank you for the careful implementation of the suggestions and for the clear communication throughout the review process.
