+++ vignettes/Alaska.Rmd
@@ -1,9 +1,10 @@
 ---
 title: "Alaska"
-subtitle: "NDVI analysis {remotePARTS}"
-author: "Clay Morrow"
-date: "`r Sys.Date()`"
-output: rmarkdown::html_vignette
+subtitle: "NDVI analysis (remotePARTS)"
+output: 
+  rmarkdown::html_vignette:
+    highlight: tango
+    df_print: tibble
 vignette: >
   %\VignetteIndexEntry{Alaska}
   %\VignetteEngine{knitr::rmarkdown}
@@ -18,26 +19,33 @@ knitr::opts_chunk$set(collapse = TRUE, comment = "##", dpi = 50)
 
 # Introduction 
 
-This vignette will demonstrate the main functionality of `remotePARTS` by working
-through a real remote sensing data set. The example follows the more-detailed
-development of the methods in Ives et al. (2021, Remote Sensing of Environment, 
-https://doi.org/10.1016/j.rse.2021.112678).
+This vignette will demonstrate the main functionality of `remotePARTS` by 
+working through a real remote sensing data set. The example follows the 
+more detailed development of the methods in Ives et al. (2021, Remote Sensing 
+of Environment, https://doi.org/10.1016/j.rse.2021.112678).
 
-First, install/update `remotePARTS` from github if needed:  
+First, install `remotePARTS`:  
 
+```{r installation, eval = FALSE}
+install.packages("remotePARTS")
 ```
+
+or 
+
+```{r alt_install, eval = FALSE}
 devtools::install_github("morrowcj/remotePARTS")
 ```
 
-Then, ensure that the package is loaded into your library:
+Then, load the package.
 
 ```{r load_remotePARTS}
 library(remotePARTS)
 ```
 
-This vignette will use `dplyr` and `ggplot2` for visualizing the data:
+This vignette will use a few `tidyverse` packages for visualizing the data:
 
 ```{r load_tidyverse}
+library(tidyr)
 library(dplyr)
 library(ggplot2)
 ```
@@ -55,67 +63,75 @@ theme_set(theme(panel.grid = element_blank(),
 
 # Alaska datasets
 
-`remotePARTS` ships with one spatial data object. This dataset contains NDVI 
+`remotePARTS` ships with a data object. This data set contains NDVI 
 values derived from NASA's MODIS satellite for the US State of Alaska. 
-The first object, `ndvi_AK10000`, due to package size limitations, this dataset is a 
-random sampling of 10,000 pixels from the full Alaska dataset. It is important 
-to note, though that `remotePARTS` can handle the full map (and much larger maps).
+The object is `ndvi_AK10000` and, due to package size limitations, this 
+data set is a random sampling of 10,000 pixels from the full Alaska data set. 
+It is important to note, though, that `remotePARTS` can handle the full map 
+(and much larger maps).
 
-For this vignette, we well also create a smaller 3000 pixel subsample `ndvi_AK3000`
-for demonstrative purposes:
+For this vignette, we will also create a smaller 3000 pixel subsample 
+`ndvi_AK3000` for demonstrative purposes:
 
 ```{r load_data}
 data("ndvi_AK10000")
-ndvi_AK3000 <- ndvi_AK10000[seq_len(3000),] # first 3000 pixels from the random 10K
+ndvi_AK3000 <- ndvi_AK10000[seq_len(3000),] # first 3000 pixels from the 10K
 ```
 
-`ndvi_AK10000` is a `data.frame` with 37 columns. `lng` and `lat` are longitude and
-latitude, respectively. `AR_coef` and `CLS_coef` are pre-calculated coefficient
-estimates of the time trends in ndvi from pixel-level time series analyses
[…]
