+++ vignettes/Alaska.Rmd
@@ -67,8 +67,8 @@ theme_set(theme(panel.grid = element_blank(),
 values derived from NASA's MODIS satellite for the US State of Alaska. 
 The object is `ndvi_AK10000` and, due to package size limitations, this 
 data set is a random sampling of 10,000 pixels from the full Alaska data set. 
-It is important to note, though, that `remotePARTS` can handle the full map 
-(and much larger maps).
+This is just for the sake of this vignette: `remotePARTS` can easily 
+handle the full map (and much larger maps).
 
 For this vignette, we will also create a smaller 3000 pixel subsample 
 `ndvi_AK3000` for demonstrative purposes:
@@ -85,7 +85,7 @@ estimates of the temporal ndvi trends, from pixel-level time series analyses
 (AR REML and conditional least squares, respectively). 
 These coefficient estimates have been standardized by the mean ndvi value 
 for the pixel over that time period;
-`land` is a factor representing land-cover classes; and
+`land` is a factor representing land cover classes; and
 the remaining 32 columns -- of the form `ndviYYYY` -- contain the NDVI values 
 from 1982 to 2013. These data already have rare land classes (those that 
 occur in less than 2% of pixels) removed and contain no missing values.
@@ -100,7 +100,7 @@ For this demonstration, we are interested in asking the following questions:
 
 1) Is NDVI in Alaska increasing over time?
 
-2) Are Alaska's NDVI time trends associated with land-cover classes? and 
+2) Are Alaska's NDVI time trends associated with land cover classes? and 
 
 3) Do Alaska's NDVI time trends differ with latitude?
 
@@ -118,7 +118,7 @@ tibble(ndvi_AK10000) |>
   labs(x = "Longitude", y = "Latitude", col = "ndvi", fill = "ndvi")
 ```
 
-The following figure shows how Alaska's three primary land-cover classes are 
+The following figure shows how Alaska's three primary land cover classes are 
 distributed.
 
 ```{r plot_land, fig.width = 4.5, fig.asp = .8}
@@ -232,8 +232,10 @@ ndvi_AK10000 %>%
   scale_color_viridis_c(option = "magma") +
   scale_fill_viridis_c(option = "magma") +
   guides(fill = "none") + 
-  labs(y = "Latitude", x = "Longitude", col = expression(beta[1]),
-       fill = expression(beta[1]))
+  labs(
+    y = "Latitude", x = "Longitude", col = expression(beta[1]), 
+    fill = expression(beta[1])
+  )
 ```
 
 `fitAR_map` and its conditional least-squares counterpart, `fitCLS_map`, are
@@ -260,7 +262,7 @@ D <- distm_scaled(coords)
 ```
 
 `distm_scaled()` scales distances across the spatial domain so that the
-greatest distance between two pixels is 1. Note that because `distm_scaled()`
+greatest distance between two pixels is 1. Because `distm_scaled()`
 uses `geosphere::distm()`, it treats coordinates as degrees longitude and 
 latitude on a sphere and calculates distances accordingly.
 
@@ -276,11 +278,25 @@ The function `covar_exp()` implements exponential covariance.
 
 The below figure is a semivariogram of `covar_exp` for two values of `r`.
 
-```{r visualize_range, fig.asp = 1, fig.width = 4, out.width="50%"}
-curve(covar_exp(x, r = .1), xlab = "distance", ylab = "covar_exp(distance, r)")
-curve(covar_exp(x, r = .2), add = TRUE, col = "red")
-legend("topright", legend = c("0.1", "0.2"), title = "r", 
-       col = c("black", "red"), lty = 1)
+```{r visualize_range, fig.asp = 0.9, fig.width = 3, out.width="50%"}
+ggplot() +
[…]
