### 2025-04-16 6e4ccdc
update check options
