+++ build_pipeline/check_pipeline.R
@@ -1,2 +1,0 @@
-source("build_pipeline/document_package.R")
-devtools::check()
