+++ build_pipeline/full_build_pipeline.R
@@ -1,4 +1,5 @@
 source("build_pipeline/document_package.R")
-devtools::check(build_args = c("--resave-data"))
+source("build_pipeline/rcmd_check.R")
 source("build_pipeline/build.R")
+source("build_pipeline/install.R")
 source("build_pipeline/render_vignette.R")
