### 2025-04-28 morrowcj/remotePARTS#42 @iosefa
JOSS Review Feedback - iosefa
Re: https://github.com/openjournals/joss-reviews/issues/7937#issuecomment-2776735553

## JOSS Checklist

**JOSS Review Checklists**

## Missing Items

The following is missing from the paper:

- **State of the field**: The paper should briefly compare remotePARTS with existing spatio-temporal or spatial-econometric libraries such as `sdmTMB`, `spaMM`, `R-INLA`, and the Python `PySAL/spreg` suite, so readers can see how the new partitioned GLS/SAR approach fills a gap for multi-million-pixel rasters.

## Areas of Improvement

Two minor elements that can be improved are:

- **Community Guidelines**: While the readme has a small sentence or two on how to contribute, a more general `CONTRIBUTING.md` file in the repository (maybe in the .github folder, which seems to be common in R projects) will encourage community standards and help potential contributors understand what the expected code styles and rules are (when to develop tests, etc.). You can then refer to this in your `README.md` file. Several examples can be found online. Here is one: https://github.com/tidyverse/dplyr/tree/main/.github

- **Automated Tests**:  There are automated tests in the GitHub actions file, however, these are not well described. I think that users that look at the readme should be able to see (1) how to run tests, and (2) what the state of those tests are. The README should explain how to run `devtools::test()` locally. Also, please consider using `codecov` and having a badge so that users can quickly understand testing coverage.

## Areas for Improvement and Repository Suggestions

Please consider the following recommendations to enhance user experience:

1. **Binder Integration**: Hosting the Alaska vignette on binder would help users very quickly test and understand the package. You could include a little badge with the link to this.
2. **Website**: add a website to the about section. Not to be too redundant with the suggestion above, but the link to your GitHub pages hosted documentation of the Alaska vignette (as a sort of read only view) would be quite helpful.
3. **CodeCov Badge**: Adding a code coverage badge would help document the tests.

---

I hope these help! This is a fantastic library and excellent work putting it together.
