### 2025-04-28 #7937 @iosefa
@jGaboardi @morrowcj

Sorry for the delay and great submission! I have completed my initial review and left detailed feedback via an issue on the package repo: https://github.com/morrowcj/remotePARTS/issues/42

To summarize:

I feel that the state of the field is missing from the paper. It doesn't situate remotePARTS among existing tools, like `sdmTMB`, `spdep`, `PySAL`, etc. Just a short paragraph or a couple of sentences explaining the gap that `remotePARTS` fills in relation to these libraries would be helpful.

Some minor improvements that I would like to see are to (1) add a `CONTRIBUTING.md` file or make the community guidelines a little more comprehensive, and (2) better documenting the tests.

Finally, adding a few extra badges to the repo readme; like codecov badges and a binder notebook with the Alaska vignette would help make the repository more user-friendly. Also, a website link in the About sidebar to the GitHub pages site Alaska vignette would also help readers jump straight in.

Once these tweaks are addressed, I believe it will be ready for publication.
