### 2025-04-29 morrowcj/remotePARTS#42 @morrowcj
@iosefa, thank you for your review. I have created a new branch with which I will try addressing your comments. I'll link the PR once it is ready and request your review of it.

Before then, I'd like to respond to your comments and ask for clarification where relevant:

> The following is missing from the paper:
> State of the field: The paper should briefly compare remotePARTS with existing spatio-temporal or spatial-econometric libraries...

This is a fair point. The packages that you referenced are not dissimilar from remotePARTS in terms of their spatio-temporal capabilities and in some instances have more sophisticated functionality. What differentiates our package is scale: to our knowledge none of these packages - and indeed no other method - can perform these kinds of analyses at the scale made possible by the PARTS method, which is routinely represented in remote sensing (e.g., millions or billions of pixels through time). I will update the paper to better reflect this point.

> Community Guidelines:

I will look into and consider adding a `CONTRIBUTING.md` file, but I don't know how its content would differ from the "Contribution" section of the README at this stage.

> Automated Tests: There are automated tests in the GitHub actions file, however, these are not well described.
> I think that users that look at the readme should be able to see (1) how to run tests, and (2) what the state of those tests are.
> The README should explain how to run devtools::test() locally. Also, please consider using codecov and having a badge so
> that users can quickly understand testing coverage.

To clarify: what do you mean by "what the state of those tests are"? Is it common practice to tell users how to run the package tests in the README? I've never seen this before and my understanding (admittedly limited) is that they are meant to ensure that the package works as expected, especially after updates.

Our code coverage is probably relatively low: I opted to write tests for the critical components but, since I'm the only maintainer and there is no more funding for development, I am hesitant to add a new component (`codecov` and additional tests) at this time.
