### 2025-05-03 morrowcj/remotePARTS#42 @iosefa
hi @morrowcj ,

> I will update the paper to better reflect this point.

Thanks!

> but I don't know how its content would differ from the "Contribution" section of the README at this stage.

Having a CONTRIBUTING.md file is useful even if it initially overlaps with the README, because it is the standard place newcomers look for contribution rules. A dedicated file gives you more room to detail this process without making the README.md file too long. Also, GitHub will highlight a contributing.md file if its in the repo root or docs folder or .github folder when new contributors go to open a PR (see screenshot below). In your contributing.md file, you can list your coding style and linting rules, branching and PR workflow, any conventions with commit message, code of conduct, release process, dev setup, project scope and philosophy, etc. There are plenty of examples on the internet to take inspiration from.

 > To clarify: what do you mean by "what the state of those tests are"? Is it common practice to tell users how to run the package tests in the README? I've never seen this before and my understanding (admittedly limited) is that they are meant to ensure that the package works as expected, especially after updates.

Yes, many OSS projects have a quick “how to run tests” section. This is often in the README or CONTRIBUTING.md file and the goal isnt to force all users to run them, but to empower would-be contributors. Having the text that explicitly documents how to run then the tests encourages reproducibility and transparency. It also helps contributors verify that they haven't broken anything and to show at a glance that the codebase is monitored by tests.

A minimal section could be:

```R
### Running tests locally (from the cloned repository directory)
# Install dev dependencies, then:
devtools::test()      # run unit tests in tests/testthat/
devtools::check()  # full R CMD check (tests + docs)
```

> Our code coverage is probably relatively low: I opted to write tests for the critical components but, since I'm the only maintainer and there is no more funding for development, I am hesitant to add a new component (codecov and additional tests) at this time.

That's fair. You could add the snippet above without integrating codecov. Coverage badges are nice to have but are certainly not mandatory. It's only important to make it clear how to run the existing tests so outside contributors can help maintain or extend them.

Thanks,
sefa
