### 2025-05-06 6e32481
update paper with software comparison.
