+++ .github/CONTRIBUTING.md
@@ -1,0 +1,18 @@
+# Contributing to remotePARTS
+
+Here, we outline how to suggest changes to remotePARTS. 
+
+## Issues
+
+For all suggestions, including bug reports, feature requests, documentation, 
+and typo or language corrections, please submit an issue on Github. 
+
+## Code contribution
+
+Substantial changes can be suggested by submitting a pull request, provided
+that there is an associated (linked) issue. If there is not an existing issue 
+concerning the changes provided in the pull request, you should create the 
+issue first and link it in the pull request. 
+
+Ideally, pull requests should be small in scope (i.e., dedicated to a single
+issue/feature/functionality).
