### 2025-05-07 morrowcj/remotePARTS#42 @morrowcj
@iosefa, I have added all features you requested to #43 (branch: [issue_42](https://github.com/morrowcj/remotePARTS/tree/issue_42), except the Binder integration. I've created an Issue (#45) with hopes to get to it at a later date.

I sent you a request to be a contributor to the repo so that I can formally request your review on the PR, but if you'd prefer to just look at the branch and provide any additional feedback here, that's fine.

Thanks for all your comments and time - they've helped improved the project a lot.

----
☁

🐋 🌻

☁
