### 2025-05-08 a36c3a8
began example script
