+++ R/_example_combineGLS.R
@@ -1,175 +1,0 @@
-# ---- Load packages ----
-
-library(tidyverse)
-library(remotePARTS)
-
-# ---- Setup and parameters ----
-
-# load the example data
-data(ndvi_AK10000)
-
-# create partition matrix, with 20 partitions
-pm = sample_partitions(npix = nrow(ndvi_AK10000), npart = 20)
-
-# specify the number of pairs that will be considered for the cross-partition analysis
-ncross = 10
-
-# specify where the RDS objects will get saved
-save_dir = "data-raw/GLS_part_obs"
-
-# should results be overwritten if they already exist?
-overwrite = TRUE
-
-# number of cores to use
-ncores = 6
-
-# ---- Fit and save the GLS ----
-
-# create the directory if it doesn't exist
-if(!dir.exists(save_dir)) {
-  dir.create(save_dir)
-}
-
-# create a counter for the number of GLS actually fit
-n_GLS_fitted = 0
-
-# fit a GLS object for each partition
-for (i in seq_len(ncol(pm))) {
-  # convert i into a fixed-length string, with padded zeros
-  i_str = str_pad(i, 2, side = "left", pad = "0")
-
-  # should the xx and invchol be saved for this iteration?
-  save_extra = (i <= ncross)
-
-  # convert to a string to identify the files with these
-  xx_str = ifelse(save_extra, "_with-xx", "")
-
-  # build the file name for this partition
-  name_i = paste0("AK10000_GLS_results_partition", i_str, xx_str, ".rds")
-
-  # and the full path
-  file_i = file.path(save_dir, name_i)
-
-  # skip to partition if this file already exists (unless overwrite=TRUE)
-  if (file.exists(file_i) & !overwrite) {
-    next
-  }
-
-
-  # get the column of partition matrix to use
-  part_i = pm[, i]
-
-  # subset the data by the partition
-  data_i <- ndvi_AK10000[part_i, ]
-
-  # get/set parameter values (set to constants for ease here)
-  nug = 0
-  rng = 0.01
-
-  # get the coordinates for this subset
-  coords_i = data_i[, c("lng", "lat")]
-
-  # calculate the varcov matrix
-  V = covar_exp(distm_km(coords_i), range = rng)
-
-  # fit the GLS - note that only the first 10 (ncross) will have xx and invchol
-  GLS_i <- fitGLS(
-    CLS_coef ~ lat + land, data = data_i, V = V, nugget = nug,
-    save.xx = save_extra, save.invchol = save_extra, ncores = ncores
-  )
-
-  # Add coordinates to the GLS for the partitions to cross
-  if (save_extra){
-    GLS_i$coords <- coords_i
-  }
-  ## TODO: this should be added to any function that automates this.
-
-  # save the RDS to a file
-  saveRDS(GLS_i, file = file_i)
-
-  # increment the iterator
-  n_GLS_fitted = n_GLS_fitted + 1
-}
-
-# ---- Calculate cross-partition statistics ----
-
-# get a list of all the partitioned results that have xx
-file_list = list.files(save_dir, pattern = "with-xx", full.names = TRUE)
-
-# calculate the number of partitions to cross
-ncross = length(file_list)
-
-# get a table of all the possible crosses
-cross.pairs = t(combn(seq_len(ncross), 2))
-
-# indicate when a new GLS needs to be loaded for a given pair
[…]
