### 2025-05-10 morrowcj/remotePARTS#42 @iosefa
> [@iosefa](https://github.com/iosefa), I have added all features you requested to [#43](https://github.com/morrowcj/remotePARTS/pull/43) (branch: [issue_42](https://github.com/morrowcj/remotePARTS/tree/issue_42)), except the Binder integration. I've created an Issue ([#45](https://github.com/morrowcj/remotePARTS/issues/45)) with hopes to get to it at a later date.
>
> I sent you a request to be a contributor to the repo so that I can formally request your review on the PR, but if you'd prefer to just look at the branch and provide any additional feedback here, that's fine.
>
> Thanks for all your comments and time - they've helped improved the project a lot.
>
> ☁
>
> 🐋 🌻
>
> ☁

@morrowcj I’ve accepted the invitation, submitted my review, and approved the PR.

The new paragraph comparing remotePARTS to existing tools satisfies the JOSS *State of the Field* criterion, and the additional polish (Codecov badge, CONTRIBUTING, etc.) all looks good. Everything is ready to merge from my end—thanks for promptly addressing the review items.
