### 2025-05-10 #7937 @iosefa
> [@iosefa](https://github.com/iosefa) -- Send me a ping here once you go over the updates from [@morrowcj](https://github.com/morrowcj) and give feedback.

@jGaboardi I’ve reviewed the updates in PR #43 and approved it. Everything now meets the JOSS requirements from my checklist.

@morrowcj Thanks for addressing the revisions so promptly—looks great!

I recommend the submission proceed to acceptance.
