### 2025-05-10 bb4c767
Update README.Rmd
fixed typo
