### 2025-05-10 #7937 @morrowcj
I have merged [PR 43](https://github.com/morrowcj/remotePARTS/pull/43), which means that the `joss_submission` branch has all the reviewer suggested changes.

Thanks again, @iosefa, @jGaboardi, @karlosmantilla!
