### 2025-05-10 #7937 @jGaboardi
@morrowcj – Can you add the DOI that [resolves the CRAN site](https://cran.r-project.org/web/packages/remotePARTS/) – `10.32614/cran.package.remoteparts` – to follow reference:

> * Ives, A. R., & Morrow, C. J. (2023). remotePARTS: Spatiotemporal autoregression analyses for large data sets. In CRAN repository. cran. https://cran.r-project.org/web/packages/remotePARTS/index.html
