### 2025-05-11 fe1ffe8
update NEWS, JOSS
