+++ paper.bib
@@ -40,7 +40,8 @@
   month = sep,
   publisher = CRAN,
   journal = {CRAN repository},
-  url = {https://cran.r-project.org/web/packages/remotePARTS/index.html}
+  doi = {10.32614/CRAN.package.remotePARTS}
+  url = {https://cran.r-project.org/package=remotePARTS}
 }
 
 @misc{alaska,
