+++ README.Rmd
@@ -38,7 +38,7 @@ done by partitioning the data into smaller chunks, analyzing chunks separately
 and then combining the separate analyses into a single test, that accounts for
 correlations among chunks, of the map-scale hypotheses.
 
-## Instalation
+## Installation
 
 To install the package and it's dependencies from CRAN, use the following R code:
 
@@ -121,6 +121,9 @@ small spatiotemporal data set for analysis:
 ```{r, warning=FALSE, message=FALSE}
 library(tibble); library(dplyr); library(tidyr); library(viridisLite)
 library(ggplot2); library(remotePARTS)
+
+# set a random seed, for reproducibility
+set.seed(42) # don't panic
 
 # simulate a spatiotemporal response variable
 sim_spatiotemp <- function(
