+++ .github/workflows/test-coverage.yaml
@@ -2,7 +2,7 @@
 # Need help debugging build failures? Start at https://github.com/r-lib/actions#where-to-find-help
 on:
   push:
-    branches: [main, master, develop]
+    branches: [main, master, develop, joss_submission]
   pull_request:
 
 name: test-coverage.yaml
