+++ paper.bib
@@ -40,7 +40,7 @@
   month = sep,
   publisher = CRAN,
   journal = {CRAN repository},
-  doi = {10.32614/CRAN.package.remotePARTS}
+  doi = {10.32614/CRAN.package.remotePARTS},
   url = {https://cran.r-project.org/package=remotePARTS}
 }
 
