+++ paper.bib
@@ -29,7 +29,7 @@
   volume = 29,
   issue = 16,
   url = {https://onlinelibrary.wiley.com/doi/10.1111/gcb.16800},
-  doi = {doi/10.1111/gcb.16800}
+  doi = {10.1111/gcb.16800}
 }
 
 
