+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: remotePARTS
 Title: Spatiotemporal Autoregression Analyses for Large Data Sets
-Version: 1.0.4
+Version: 1.1
 Authors@R:
     c(person(given = "Clay",
            family = "Morrow",
@@ -52,5 +52,4 @@ Suggests:
     sf,
     testthat (>= 3.0.0)
 LinkingTo: Rcpp, RcppEigen
-VignetteBuilder: knitr
 Config/testthat/edition: 3
