### 2025-05-12 #7937 @morrowcj
> please review the Additional Author Tasks After Review is Complete https://github.com/openjournals/joss-reviews/issues/7937#issuecomment-2869084711 for those 5 bullet points.

@jGaboardi, I believe I have completed all tasks.

> Can you add the DOI that resolves the CRAN site

I updated `paper.bib` to include the DOI in commit [fe1ffe8](https://github.com/morrowcj/remotePARTS/commit/fe1ffe8f80fb1e5bf26f72ac340dea7c2699c503)
