### 2025-05-12 #7937 @jGaboardi
@morrowcj Looks like we made a minor oversight in the DOI for the (Lewińska et al) paper. Can you correct that? Simply remove the extra "doi":

`https://doi.org/*doi*/10.1111/gcb.16800`
