### 2025-05-12 #7937 @editorialbot
:wave: @openjournals/ese-eics, this paper is ready to be accepted and published.
